# server/openapi.yaml
openapi: 3.1.0
...
info:
  title: MetanoiaDocs API
  version: '0.1'
  description: |
    Everything here is same-origin with the app. Two ways to authenticate:

    - **A session cookie** (`md_session`), which is what the browser uses.
    - **A personal access token** as `Authorization: Bearer <token>`, which is
      what everything else uses. Make one in Settings → API tokens. A token
      carries exactly the access of the person who made it — there are no
      scopes, because there is nothing a token should be able to do that its
      owner cannot.

    Times are ISO-8601. Dates on tasks are `YYYY-MM-DD` — a due date is a day,
    not an instant. Errors are `{ "error": "<sentence>" }` with a 4xx or 5xx
    status; the sentence is written to be shown to a person.
  license:
    name: MIT
    url: https://github.com/Xephyr-Labs/metanoiadocs/blob/main/LICENSE
servers:
  - url: http://localhost:8092
    description: A local instance from the quick start
  - url: https://docs.example.com
    description: Your instance
security:
  - bearerAuth: []
...
paths:
...
  /api/webhooks:
    get:
      tags: [Webhooks]
      summary: Subscriptions
      description: Admin only.
      responses:
        '200':
          description: Hooks, without their secrets
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Webhook' } }
        '403': { $ref: '#/components/responses/Error' }
    post:
      tags: [Webhooks]
      summary: Subscribe a URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri }
                events:
                  type: array
                  items: { type: string }
                  description: Empty means every event.
      responses:
        '200':
          description: The hook, with the one and only copy of its secret
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Webhook' }
                  - type: object
                    properties: { secret: { type: string } }
...
  /api/webhooks/{id}:
    patch:
      tags: [Webhooks]
      summary: Change a hook
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string, format: uri }
                events: { type: array, items: { type: string } }
                active: { type: boolean }
      responses:
        '200':
          description: The hook
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Webhook' }
    delete:
      tags: [Webhooks]
      summary: Delete a hook
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /api/webhooks/{id}/rotate:
    post:
      tags: [Webhooks]
      summary: Issue a new signing secret
      description: The old one stops verifying immediately.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: The hook and its new secret
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Webhook' }
                  - type: object
                    properties: { secret: { type: string } }
  /api/webhooks/{id}/test:
    post:
      tags: [Webhooks]
      summary: Send a `ping` now and say what came back
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: The outcome
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  statusCode: { type: [integer, 'null'] }
                  attempts: { type: integer }
                  error: { type: [string, 'null'] }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A personal access token from Settings → API tokens.
    cookieAuth:
      type: apiKey
      in: cookie
      name: md_session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Ok:
      description: Done
      content:
        application/json:
          schema: { type: object, properties: { ok: { const: true } } }
    Error:
      description: A sentence saying what went wrong
      content:
        application/json:
          schema:
            type: object
            properties: { error: { type: string } }
  schemas:
...
    Webhook:
      type: object
      properties:
        id: { type: string }
        url: { type: string, format: uri }
        events: { type: array, items: { type: string }, description: Empty means all }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
        last_ok: { type: [boolean, 'null'] }
        last_at: { type: [string, 'null'], format: date-time }
...